drow_ai_backend/requirements.txt
numpy<2
pillow
opencv-python
peft 
diffusers 
transformers 
scikit-learn 
ftfy 
accelerate 
invisible_watermark 
safetensors 
controlnet-aux 
mediapipe 
timm 
GPUtil
fastapi 
nest-asyncio 
uvicorn 
pyngrok
